src/fastapi_app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
redis==4.6.0
paho-mqtt==2.1.0
feast[redis]==0.40.0
pandas==2.2.3
